+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Create `Composite` class, which `CompositeArgs` now subclasses.
+
 ## [0.8.0] - 2024-10-29
 
 ### Fixed